feat: Enhance libop with improved CMake configuration and memory management
- Added OP_VERSION definition in CMakeLists.txt.
- Updated linking for MinHook to use CMake package if available.
- Refactored libop to use std::atomic for s_id and std::unique_ptr for m_context.
- Replaced raw pointers with std::vector for screen data management in libop.

+++ libop/include/promutex.h
@@ -41,7 +41,7 @@ class promutex {
     }
 
     DWORD try_lock(size_t time_) {
-        return ::WaitForSingleObject(_hmutex, time_);
+        return ::WaitForSingleObject(_hmutex, static_cast<DWORD>(time_));
     }
     void unlock() {
         assert(_hmutex);
